refactor to have processor details state, validation state, property state, and transform state separated in the store
