+++ nifi-frontend/src/main/frontend/apps/nifi-jolt-transform-ui/src/app/pages/jolt-transform-json-ui/feature/jolt-transform-json-ui.module.ts
@@ -23,7 +23,7 @@ import { StoreModule } from '@ngrx/store';
 import { joltTransformJsonUiFeatureKey, reducers } from '../state';
 import { EffectsModule } from '@ngrx/effects';
 import { MatDialogModule } from '@angular/material/dialog';
-import { JoltTransformJsonUiEffects } from '../state/jolt-transform-json-ui/jolt-transform-json-ui.effects';
+import { JoltTransformJsonProcessorDetailsEffects } from '../state/jolt-transform-json-processor-details/jolt-transform-json-processor-details.effects';
 import { CodemirrorModule } from '@ctrl/ngx-codemirror';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatCardModule } from '@angular/material/card';
@@ -44,7 +44,7 @@ import { MatExpansionModule } from '@angular/material/expansion';
         CommonModule,
         JoltTransformJsonUiRoutingModule,
         StoreModule.forFeature(joltTransformJsonUiFeatureKey, reducers),
-        EffectsModule.forFeature(JoltTransformJsonUiEffects),
+        EffectsModule.forFeature(JoltTransformJsonProcessorDetailsEffects),
         MatDialogModule,
         CodemirrorModule,
         ReactiveFormsModule,
