+++ nifi-frontend/src/main/frontend/apps/nifi-jolt-transform-ui/src/app/pages/jolt-transform-json-ui/state/jolt-transform-json-processor-details/jolt-transform-json-processor-details.actions.ts
@@ -1,0 +1,41 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { createAction, props } from '@ngrx/store';
+import { ProcessorDetails } from './index';
+import { HttpErrorResponse } from '@angular/common/http';
+
+const JOLT_TRANSFORM_JSON_PROCESSOR_DETAILS_PREFIX = '[Jolt Transform Json Processor Details]';
+
+export const resetJoltTransformJsonProcessorDetailsState = createAction(
+    `${JOLT_TRANSFORM_JSON_PROCESSOR_DETAILS_PREFIX} Reset Jolt Transform Json Processor Details State`
+);
+
+export const loadProcessorDetails = createAction(
+    `${JOLT_TRANSFORM_JSON_PROCESSOR_DETAILS_PREFIX} Load Processor Details`,
+    props<{ processorId: string }>()
+);
+
+export const loadProcessorDetailsSuccess = createAction(
+    `${JOLT_TRANSFORM_JSON_PROCESSOR_DETAILS_PREFIX} Load Processor Details Success`,
+    props<{ response: ProcessorDetails }>()
+);
+
+export const loadProcessorDetailsFailure = createAction(
+    `${JOLT_TRANSFORM_JSON_PROCESSOR_DETAILS_PREFIX} Load Processor Details Failure`,
+    props<{ response: HttpErrorResponse }>()
+);
