+++ nifi-frontend/src/main/frontend/apps/nifi-jolt-transform-ui/src/app/pages/jolt-transform-json-ui/state/jolt-transform-json-property/jolt-transform-json-property.reducer.ts
@@ -1,0 +1,62 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { JoltTransformJsonPropertyState } from './index';
+import { createReducer, on } from '@ngrx/store';
+import {
+    resetJoltTransformJsonPropertyState,
+    resetSavePropertiesState,
+    saveProperties,
+    savePropertiesFailure,
+    savePropertiesSuccess
+} from './jolt-transform-json-property.actions';
+
+export const initialState: JoltTransformJsonPropertyState = {
+    saving: false,
+    loadedTimestamp: '',
+    status: 'pending',
+    savingProperties: false,
+    savePropertiesResponse: null,
+    savePropertiesFailureResponse: null
+};
+
+export const joltTransformJsonPropertyReducer = createReducer(
+    initialState,
+    on(resetJoltTransformJsonPropertyState, () => ({
+        ...initialState
+    })),
+    on(saveProperties, (state) => ({
+        ...state,
+        savingProperties: true
+    })),
+    on(savePropertiesSuccess, (state, { response }) => ({
+        ...state,
+        savingProperties: false,
+        savePropertiesResponse: response
+    })),
+    on(savePropertiesFailure, (state, { response }) => ({
+        ...state,
+        savingProperties: false,
+        savePropertiesFailureResponse: response
+    })),
+    on(resetSavePropertiesState, (state) => ({
+        ...state,
+        savingProperties: false,
+        savePropertiesResponse: null,
+        savePropertiesFailureResponse: null
+    }))
+);
