+++ nifi-frontend/src/main/frontend/apps/nifi-jolt-transform-ui/src/app/pages/jolt-transform-json-ui/state/jolt-transform-json-transform/jolt-transform-json-transform.reducer.ts
@@ -1,0 +1,56 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { JoltTransformJsonTransformState } from './index';
+import { createReducer, on } from '@ngrx/store';
+import {
+    resetJoltTransformJsonTransformState,
+    transformJoltSpec,
+    transformJoltSpecFailure,
+    transformJoltSpecSuccess
+} from './jolt-transform-json-transform.actions';
+
+export const initialState: JoltTransformJsonTransformState = {
+    saving: false,
+    loadedTimestamp: '',
+    status: 'pending',
+    transformingJoltSpec: false,
+    transformationResponse: null,
+    transformationFailureResponse: null
+};
+
+export const joltTransformJsonTransformReducer = createReducer(
+    initialState,
+    on(resetJoltTransformJsonTransformState, () => ({
+        ...initialState
+    })),
+    on(transformJoltSpec, (state) => ({
+        ...state,
+        transformingJoltSpec: true
+    })),
+    on(transformJoltSpecSuccess, (state, { response }) => ({
+        ...state,
+        transformingJoltSpec: false,
+        transformationResponse: response
+    })),
+    on(transformJoltSpecFailure, (state, { response }) => ({
+        ...state,
+        transformingJoltSpec: false,
+        transformationResponse: null,
+        transformationFailureResponse: response
+    }))
+);
