+++ nifi-frontend/src/main/frontend/apps/nifi-jolt-transform-ui/src/app/pages/jolt-transform-json-ui/state/jolt-transform-json-validate/jolt-transform-json-validate.reducer.ts
@@ -1,0 +1,62 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { JoltTransformJsonValidateState } from './index';
+import { createReducer, on } from '@ngrx/store';
+import {
+    resetJoltTransformJsonValidateState,
+    resetValidateJoltSpecState,
+    validateJoltSpec,
+    validateJoltSpecFailure,
+    validateJoltSpecSuccess
+} from './jolt-transform-json-validate.actions';
+
+export const initialState: JoltTransformJsonValidateState = {
+    saving: false,
+    loadedTimestamp: '',
+    status: 'pending',
+    validatingJoltSpec: false,
+    validationResponse: null,
+    validationFailureResponse: null
+};
+
+export const joltTransformJsonValidateReducer = createReducer(
+    initialState,
+    on(resetJoltTransformJsonValidateState, () => ({
+        ...initialState
+    })),
+    on(validateJoltSpec, (state) => ({
+        ...state,
+        validatingJoltSpec: true
+    })),
+    on(validateJoltSpecSuccess, (state, { response }) => ({
+        ...state,
+        validatingJoltSpec: false,
+        validationResponse: response
+    })),
+    on(validateJoltSpecFailure, (state, { response }) => ({
+        ...state,
+        validatingJoltSpec: false,
+        validationFailureResponse: response
+    })),
+    on(resetValidateJoltSpecState, (state) => ({
+        ...state,
+        validatingJoltSpec: null,
+        validationResponse: null,
+        validationFailureResponse: null
+    }))
+);
